+++ public/hud.js
@@ -15,8 +15,8 @@ function fillHUD(player, report) {
 
     player.querySelector("#playerName").innerText = report.playerName;
 
-    player.querySelector("#tc .value").innerText = report.clicks;
-    player.querySelector("#apm .value").innerText = `${report.apm} (${report.maxApm})`;
+    player.querySelector("#tc .value").innerText = report.actions;
+    player.querySelector("#apm .value").innerText = `${report.apm}/${report.cleanApm} (${report.maxApm})`;
     player.querySelector("#tmp .value").innerText = report.totalMPSpent;
     player.querySelector("#av .value").innerText = report.armyValue;